long to bigint

+++ src/main/java/TOTPGenerator.java
@@ -2,7 +2,9 @@ import helpers.URIHelper;
 import interfaces.ITOTPGenerator;
 
 import java.io.UnsupportedEncodingException;
+import java.math.BigInteger;
 import java.net.URI;
+import java.net.URISyntaxException;
 import java.time.Duration;
 import java.time.Instant;
 import java.util.Date;
@@ -152,7 +154,7 @@ public class TOTPGenerator extends OneTimePasswordGenerator implements ITOTPGene
      * @throws IllegalArgumentException when code could not be generated
      */
     @Override
-    public String generate(long secondsPast1970) throws IllegalArgumentException {
+    public String generate(BigInteger secondsPast1970) throws IllegalArgumentException {
         if (!validateTime(secondsPast1970)) {
             throw new IllegalArgumentException("Time must be above zero");
         }
@@ -183,7 +185,7 @@ public class TOTPGenerator extends OneTimePasswordGenerator implements ITOTPGene
      * @return counter based on a specific date and time interval
      */
     private long calculateCounter(Date date, Duration period) {
-        return calculateCounter(date.getTime(), period);
+        return calculateCounter(BigInteger.valueOf(date.getTime()), period);
     }
 
     /**
@@ -192,8 +194,8 @@ public class TOTPGenerator extends OneTimePasswordGenerator implements ITOTPGene
      * @param period time interval between new codes
      * @return counter based on seconds past 1970 and time interval
      */
-    private long calculateCounter(long secondsPast1970, Duration period) {
-        return TimeUnit.SECONDS.toMillis(secondsPast1970) / TimeUnit.SECONDS.toMillis(period.getSeconds());
+    private long calculateCounter(BigInteger secondsPast1970, Duration period) {
+        return TimeUnit.SECONDS.toMillis(secondsPast1970.intValue()) / TimeUnit.SECONDS.toMillis(period.getSeconds());
     }
 
     /**
@@ -210,7 +212,7 @@ public class TOTPGenerator extends OneTimePasswordGenerator implements ITOTPGene
      * @param time time value to check against
      * @return whether time is above zero
      */
-    private boolean validateTime(long time) {
-        return time > 0;
+    private boolean validateTime(BigInteger time) {
+        return time.intValue() > 0;
     }
 }
