+++ src/main/java/interfaces/ITOTPGenerator.java
@@ -1,5 +1,6 @@
 package interfaces;
 
+import java.math.BigInteger;
 import java.time.Duration;
 import java.time.Instant;
 import java.util.Date;
@@ -8,7 +9,7 @@ public interface ITOTPGenerator {
     String generate();
     String generate(Instant instant);
     String generate(Date date);
-    String generate(long secondsPast1970);
+    String generate(BigInteger secondsPast1970);
     boolean verify(String code);
     Duration getPeriod();
 }
